Fixed bug with course-to-fine
I was down-sampling the down-samples after each iteration instead of down-sampling the original point clouds.
Now if the step size is too small, dT is shrunk to 0.75 to see if we can get dT and dR to converge.
Now If the norm of dT and dR is not small enough, the transformation is discarded. This eps_2 was increased